requirements.txt
altgraph==0.17.4
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.62
botocore==1.34.62
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
distlib==0.3.8
Django==5.0.3
django-storages==1.14.2
filelock==3.13.1
idna==3.6
jmespath==1.0.1
packaging==24.0
pefile==2023.2.7
platformdirs==4.2.0
psutil==5.9.8
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.31.0
s3transfer==0.10.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
virtualenv==20.25.1
